Refactored Authenticator and added two methods which delegate to PasswordManager so that Publisher can use Authenticator instead of PasswordManager.
- Authenticator and its tests got moved into utils.
- No longer retrieve "repositories.{name}.url" in two steps ("repositories.{name}", then "url") since that does not allow us to use environment variable POETRY_REPOSITORIES_{name}_URL.
